# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EchoID Nexus API
  description: |
    AI-powered academic assistant for UIT University.
    This document is the **single source of truth** for all API contracts.
    No endpoint is implemented before it appears here.
  version: 0.1.0
  contact:
    name: Team Neurorise
    email: admin@uit.ac.ma
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://backend:8080
    description: Docker internal
...
paths:
...
  /api/lms/moodle/courses:
    get:
      tags: [LMS]
      summary: List Moodle courses
      operationId: listMoodleCourses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Course list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseList'
  /api/lms/moodle/courses/{id}:
    get:
      tags: [LMS]
      summary: Moodle course detail with section tree
      operationId: getMoodleCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseDetailDto'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/lms/gclassroom/courses:
    get:
      tags: [LMS]
      summary: List Google Classroom courses
      operationId: listGClassroomCourses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Course list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseList'
  /api/lms/gclassroom/courses/{id}:
    get:
      tags: [LMS]
      summary: Google Classroom course detail
      operationId: getGClassroomCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseDetailDto'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Events ──────────────────────────────────────
...
  /api/students/{id}/workload:
    get:
      tags: [Workload]
      summary: Get student workload density (Wd)
      operationId: getStudentWorkload
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workload response returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWdResponseDto'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/{id}/device-token:
    post:
      tags: [Workload]
      summary: Register push notification device token
      operationId: registerDeviceToken
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenRequest'
      responses:
        '200':
          description: Token registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Notifications ───────────────────────────────
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      operationId: listNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotificationList'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not your notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Admin ───────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeviceTokenRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
          example: "ExponentPushToken[xxxxxxxxxxxxxx]"
        platform:
          type: string
          enum: [IOS, ANDROID]

    # ── Response DTOs ─────────────────────────────
...
    CourseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Introduction to Machine Learning
        lmsSource:
          type: string
          enum: [MOODLE, GOOGLE_CLASSROOM]
        school:
          type: string
          enum: [ENSA, EST, FAC, OTHER]
        semester:
          type: string
          enum: [S1, S2, ANNUAL]
        isActive:
          type: boolean
    CourseDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        lmsSource:
          type: string
          enum: [MOODLE, GOOGLE_CLASSROOM]
        school:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/CourseSectionDto'
    CourseSectionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Chapter 3: Sorting Algorithms"
        orderIndex:
          type: integer
...
    WdResponseDto:
      type: object
      properties:
        wdScore:
          type: number
          format: double
          example: 3.75
        status:
          type: string
          enum: [LOW, MODERATE, HIGH, CRITICAL]
        breakdown:
          type: array
          items:
            $ref: '#/components/schemas/WdBreakdownItem'
        history:
          type: array
          items:
            $ref: '#/components/schemas/WdHistoryPoint'
        calculatedAt:
          type: string
          format: date-time
    WdBreakdownItem:
      type: object
      properties:
        courseTitle:
          type: string
        assignmentTitle:
          type: string
        ci:
          type: number
          format: double
        ti:
          type: number
          format: double
        contribution:
          type: number
          format: double
    WdHistoryPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        wdScore:
          type: number
          format: double
    NotificationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [DEADLINE_REMINDER, MILESTONE_NUDGE, WORKLOAD_ALERT, SYSTEM]
        channel:
          type: string
          enum: [PUSH, EMAIL]
        message:
          type: string
        isRead:
          type: boolean
        sentAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
        assignmentId:
          type: string
          format: uuid
          nullable: true
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
...
    ApiResponseCourseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
        message:
          type: string
    ApiResponseCourseDetailDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/CourseDetailDto'
        message:
          type: string
...
    ApiResponseWdResponseDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/WdResponseDto'
        message:
          type: string
    ApiResponseNotificationList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/NotificationDto'
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
        path:
          type: string